+++ NEWS.md
@@ -1,3 +1,8 @@
+# Version (development version)
+
+ * ...
+
+
 # Version 0.15.1 [2024-11-20]
 
 ## Bug Fixes